Switched to hydra config, name refactoring, cleanup
