+++ WasteAnnotator/config/community/infomap.yaml
@@ -1,0 +1,2 @@
+# @package community.infomap
+function: cdlib.algorithms.infomap
