+++ WasteAnnotator/config/community/louvain.yaml
@@ -1,0 +1,2 @@
+# @package community.louvain
+function: cdlib.algorithms.louvain
