+++ WasteAnnotator/config/envirnonment/docker.yaml
@@ -1,0 +1,7 @@
+# @package _global_
+base_path: /waste-annotator
+data_path: ${base_path}/data
+raw_data_path: ${data_path}/raw
+repository_path: ${raw_data_path}/repository
+test_data_path: ${data_path}/test
+out_path: ${data_path}
