+++ WasteAnnotator/config/exporter/default.yaml
@@ -1,0 +1,5 @@
+# @package exporter
+defaults:
+  - json
+
+force_run: false
