+++ WasteAnnotator/config/graphextractor/arcan.yaml
@@ -1,0 +1,6 @@
+_target_: graphextractor.arcan.ArcanGraphExtractor
+arcan_path: ${base_path}/src/arcan
+repository_path: ${repository_path}
+arcan_out: ${out_path}
+logs_path: ${out_path}/logs
+force_run: False
