--- WasteAnnotator/outfile
@@ -1,19 +1,0 @@
-10:00:04 [main] INFO  c.a.t.AnalyseCommand - Using in-memory database.
-10:00:06 [main] INFO  c.a.d.v.GitVersionControlSystem - Active branch set to 'master' from null.
-10:00:06 [main] INFO  c.a.a.ProjectVersioningManager - Version control system of 'JakeWharton|DiskLruCache' set to com.arcan.data.vcs.GitVersionControlSystem
-10:00:06 [main] INFO  c.a.d.v.GitVersionControlSystem - Pulling from remote github.com
-10:00:06 [main] INFO  c.a.a.ProjectVersioningManager - Discovering non-analysed versions using class com.arcan.data.ves.LastVersionEnumerationStrategy
-10:00:07 [main] INFO  c.a.a.ProjectVersioningManager - 1 non-analyzed versions discovered.
-10:00:07 [main] INFO  c.a.a.Arcan - Saved project 'JakeWharton|DiskLruCache' and '1' versions to database.
-10:00:07 [main] INFO  c.a.d.v.GitVersionControlSystem - Pulling from remote github.com
-10:00:07 [main] INFO  c.a.a.ProjectVersioningManager - Discovering non-analysed versions using class com.arcan.data.ves.LastVersionEnumerationStrategy
-10:00:07 [main] INFO  c.a.a.ProjectVersioningManager - 1 non-analyzed versions discovered.
-10:00:07 [main] INFO  c.a.a.ProjectAnalysis - Analysing 1 new versions. Total versions in JakeWharton|DiskLruCache: 1
-10:00:07 [main] INFO  c.a.a.ProjectAnalysis - Analysing version '1-3e016356' (1/1)
-10:00:07 [main] INFO  c.a.m.GraphBuilder - Source files count: 5
-10:00:08 [main] INFO  c.a.m.GraphBuilder - Model built. Scanning for dependencies.
-10:00:09 [main] INFO  c.a.m.GraphBuilder - Building the dependency graph took 2 seconds: tinkergraph[vertices:8 edges:13]
-10:00:09 [main] INFO  c.a.a.VersionAnalysis - Graph of this version: 5 units, 3 containers
-10:00:09 [main] INFO  c.a.a.ProjectAnalysis - Analysis completed.
-10:00:09 [main] INFO  c.a.a.ProjectAnalysis - Versions analysed successfully: 1/1
-10:00:09 [main] INFO  c.a.a.ProjectAnalysis - Elapsed time: 0 minutes 2 seconds
