+++ WasteAnnotator/src/annotator/__init__.py
@@ -1,0 +1,2 @@
+from .interface import Annotator
+from .autofl import AutoFLAnnotator
