--- WasteAnnotator/src/componentannotator/__init__.py
@@ -1,1 +1,0 @@
-from .componentannotator import ComponentAnnotator
