--- WasteAnnotator/src/componentextractor/__init__.py
@@ -1,2 +1,0 @@
-from .communityextractor import CommunityExtractor
-from .componentextractor import ComponentExtractor
