+++ WasteAnnotator/src/exporter/__init__.py
@@ -1,0 +1,2 @@
+from .interface import ProjectExporter
+from .json import JSONProjectExporter
