+++ WasteAnnotator/src/finder/__init__.py
@@ -1,0 +1,2 @@
+from .interface import ProjectFinder
+from .github import GitHubFinder
