+++ WasteAnnotator/src/finder/interface.py
@@ -1,0 +1,6 @@
+from typing import Protocol
+
+
+class ProjectFinder(Protocol):
+    def find_projects(self, amount: int):
+        pass
