+++ WasteAnnotator/src/graphextractor/__init__.py
@@ -1,0 +1,1 @@
+from .arcan import ArcanGraphExtractor
