+++ WasteAnnotator/src/graphextractor/interface.py
@@ -1,0 +1,8 @@
+from typing import Protocol
+
+from entities import Project
+
+
+class GraphExtractor(Protocol):
+    def extract_graph(self, project: Project) -> Project:
+        pass
