+++ WasteAnnotator/src/pipeline/complete.py
@@ -3,48 +3,49 @@ from urllib.error import HTTPError
 
 from loguru import logger
 
-from projectexporter import ProjectExporter
-from projectextractor import ProjectExtractor
+from exporter import ProjectExporter
+from finder import GitHubFinder
 
-from componentannotator import ComponentAnnotator
-from componentextractor import CommunityExtractor
-from componentextractor import ComponentExtractor
+from annotator import AutoFLAnnotator
+from community import CommunityExtractor
+from graphextractor import ArcanGraphExtractor
 
 
 class CompletePipeline:
     """
-    The ComponentAnnotator class is responsible for annotating files in abandoned GitHub projects.
-    It utilizes the ProjectExtractor to find abandoned projects, the ComponentExtractor to run the
-    Arcan tool for component information, and the auto-fl annotator for file-level annotations (weak labels).
+    The CompletePipeline class is responsible for running the complete pipeline to annotate projects.
     """
 
     def __init__(self,
-                 project_extractor: ProjectExtractor,
-                 component_extractor: ComponentExtractor,
-                 component_annotator: ComponentAnnotator,
+                 project_finder: GitHubFinder,
+                 graph_extractor: ArcanGraphExtractor,
+                 semantic_annotator: AutoFLAnnotator,
                  community_extractor: CommunityExtractor,
-                 project_exporter: List[ProjectExporter],
-                 language):
+                 project_exporter: List[ProjectExporter]
+                 ):
         """
-        Initializes the ComponentAnnotator with default values for the ProjectExtractor.
 
         Args:
-            language: The programming language used in the project.
+            project_finder:
+            graph_extractor:
+            semantic_annotator:
+            community_extractor:
+            project_exporter:
         """
-        self.project_extractor: ProjectExtractor = project_extractor
-        self.component_extractor: ComponentExtractor = component_extractor
-        self.component_annotator: ComponentAnnotator = component_annotator
+
+        self.project_finder: GitHubFinder = project_finder
+        self.graph_extractor: ArcanGraphExtractor = graph_extractor
+        self.semantic_annotator: AutoFLAnnotator = semantic_annotator
         self.community_extractor: CommunityExtractor = community_extractor
         self.project_exporter: List[ProjectExporter] = project_exporter
-        self.language: str = language
 
-        logger.info(f"Initialized ComponentAnnotator (project programming language -> {language})")
+        logger.info(f"Initialized ComponentAnnotator")
 
     def run(self, num_proj: int = 10):
         abandoned_projects = []
         try:
             logger.info("Starting to retrieve abandoned projects from GitHub")
-            abandoned_projects = self.project_extractor.find_abandoned_projects(num_proj)
+            abandoned_projects = self.project_finder.find_projects(num_proj)
             logger.info("Finished retrieving abandoned projects from GitHub")
         except HTTPError as exc:
             logger.error("Failed to retrieve abandoned projects from GitHub")
@@ -52,13 +53,13 @@ class CompletePipeline:
         for project in abandoned_projects:
 
             try:
-                logger.info(f"Starting to extract project dependency graph `{project.name}`")
-                project = self.component_extractor.dependency_graph(project)
+                logger.info(f"Starting to extract dependency graph for project `{project.name}`")
+                project = self.graph_extractor.extract_graph(project)
                 logger.info(f"Finished extracting dependency graph for project `{project.name}`")
 
-                logger.info(f"Starting to annotate components of project `{project.name}`")
-                project = self.component_annotator.annotate_project(project)
-                logger.info(f"Finished annotating components of project `{project.name}`")
+                logger.info(f"Starting to annotate project `{project.name}`")
+                project = self.semantic_annotator.annotate_project(project)
+                logger.info(f"Finished annotating project `{project.name}`")
 
                 logger.info(f"Starting to extract community information for project `{project.name}`")
                 project = self.community_extractor.extract(project)
