--- WasteAnnotator/src/projectexporter/__init__.py
@@ -1,2 +1,0 @@
-from .exporter import ProjectExporter
-from .file_exporter import JSONProjectExporter
