--- WasteAnnotator/src/projectextractor/__init__.py
@@ -1,1 +1,0 @@
-from .projectextractor import ProjectExtractor
